LayoutTests/storage/indexeddb/open-close-version-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test interleaved open/close/setVersion calls in various permutations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "open-close-version.html"

TEST: setVersion blocked on open handles
self.testname = 'test1'; self.ver = 1
'h1.open'
'h1.open.onsuccess'
'h2.open'
'h2.open.onsuccess'
'h1.setVersion'
'h2.onversionchange'
    h2 closing, but not immediately
'h1.setVersion.onblocked'
'h2.close'
'h1.setVersion.onsuccess'
'h1.setVersion.transaction-complete'
PASS blockedEventFired is true

TEST: setVersion not blocked if handle closed immediately
self.testname = 'test2'; self.ver = 1
'h1.open'
'h1.open.onsuccess'
'h2.open'
'h2.open.onsuccess'
'h1.setVersion'
'h2.onversionchange'
    h2 closing immediately
'h2.close'
'h1.setVersion.onblocked'
'h1.setVersion.onsuccess'
'h1.setVersion.transaction-complete'
NOTE: Will FAIL with extra bogus h1.setVersion.onblocked step; https://bugs.webkit.org/show_bug.cgi?id=71130
FAIL blockedEventFired should be false. Was true.

TEST: open and setVersion blocked if a VERSION_CHANGE transaction is running - close when blocked
self.testname = 'test3'; self.ver = 1
'h1.open'
'h1.open.onsuccess'
'h2.open'
'h2.open.onsuccess'
'h1.setVersion'
'h2.setVersion'
'h2.onversionchange'
'h1.setVersion.onblocked'
'h1.onversionchange'
'h2.setVersion.onblocked'
    h2 blocked so closing
'h2.close'
'h3.open'
'h2.setVersion.onerror'
'h1.setVersion.onsuccess'
'h1.setVersion.transaction-complete'
'h3.open.onsuccess'
PASS blockedEventFired is true
PASS versionChangeComplete is true

TEST: open and setVersion blocked if a VERSION_CHANGE transaction is running - just close
self.testname = 'test4'; self.ver = 1
'h1.open'
'h1.open.onsuccess'
'h2.open'
'h2.open.onsuccess'
'h1.setVersion'
'h3.open'
'h2.close'
'h1.setVersion.onblocked'
'h1.setVersion.onsuccess'
'h1.setVersion.transaction-complete'
'h3.open.onsuccess'
NOTE: Will FAIL with extra bogus h1.setVersion.onblocked step; https://bugs.webkit.org/show_bug.cgi?id=71130
FAIL blockedEventFired should be false. Was true.
PASS versionChangeComplete is true

TEST: open blocked if a VERSION_CHANGE transaction is running
self.testname = 'test5'; self.ver = 1
'h1.open'
'h1.open.onsuccess'
'h1.setVersion'
'h2.open'
'h1.setVersion.onsuccess'
'h1.setVersion.transaction-complete'
'h2.open.onsuccess'
PASS versionChangeComplete is true

TEST: two setVersions from the same connection
self.testname = 'test6'; self.ver = 1
'h1.open'
'h1.open.onsuccess'
'h1.setVersion'
'h1.setVersion'
'h1.setVersion.onsuccess'
'h1.setVersion.transaction-complete'
half done
'h1.setVersion.onsuccess'
'h1.setVersion.transaction-complete'
PASS versionChangeComplete is true
PASS successfullyParsed is true

TEST COMPLETE
